Remove guard around opening i18n.json
This operation should never fail because the i18n.json file is
shipped with the package and always present.

Possibly the rationale of needing it for tests no longer applies
since the path is now absolute and not relative to the current
working directory.

See <https://github.com/freedomofpress/securedrop/pull/7443#discussion_r1963825620>.