refactor: generalize "/index" and "/metadata" to flat collections
Once the index loses the source/collection hierarchy, the client has no
way to "hint" the server in what source's collection to find a given
item.  There are other index structures that could provide this hint,
but it's simpler just to scan the Submission and Reply tables directly
(without any relationships to eager-load) for the set of item UUIDs
requested by the client.

As a bonus, the dataclasses provide sufficient structural validation of
incoming requests.