Support importing of fixed datasets
loaddata.py generates a database with randomized data, which makes it
difficult to write tests against. The new loadfixedata.py takes a
dataset that's declared in a YAML file, invoke it via:

  USE_FIXED_DATA=data.yaml make dev

The primary intent for this is to be used with the app.

Because it does override some data the script is not as efficient as it
could be, for example create_source_user() generates a PGP keypair for
the source, and then it's immediately overwritten by the keys supplied
in the YAML.

A --save-items flag can be hacked into bin/dev-deps to save encrypted
attachments so they can be reused in future runs, as otherwise the
ciphertext intentionally isn't 100% reproducible.

Not every field is supported yet, the intent is that basically
everything that would be exposed in APIv2 would be taken from the YAML.

There is a bit of logic added to dev-shell to ensure that even if the
YAML file is in some other directory (e.g. the app), it still gets
mounted into the container so the script can read from it.