Set noble end-of-life date
Historically we only did this closer to the end of life, but that left
very old SecureDrops that were on an outdated version still up because
they never picked up the code update that shipped the EOL check.

So let's enable it now, so that regardless of whether they continue to
receive further updates, they will go down in April 2029.