feat(api2): handle "source_conversation_truncated" event
A "source_conversation_truncated" event involves deleting all the items
in the source's collection with interaction counts less than or equal to
the specified upper bound, assumed to be the last item known to the
client.  This achieves the same consistency as a
"source_conversation_deleted" event without requiring its strict
versioning.